requirements.txt
torch==2.7.1
transformers==4.53.1
accelerate==1.8.1
peft==0.16.0
trl==0.19.1
bitsandbytes==0.46.1
vllm==0.9.2
huggingface-hub==0.33.2
datasets==4.0.0
scikit-learn==1.7.0
numpy==2.3.1
rouge_score==0.1.2
json_repair==0.47.6
openai==1.90.0
boto3==1.39.4
botocore==1.39.4
python-dotenv==1.1.1